% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_plate.R
\name{is_gr_plate}
\alias{is_gr_plate}
\title{Test whether an object is a \code{gr_plate}}
\usage{
is_gr_plate(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{gr_plate}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a \code{gr_plate}
}
\examples{
path <- system.file("extdata", "growth_long.csv", package = "gRate")
is_gr_plate(gr_read(path))
is_gr_plate(mtcars)
}
